Module ledge_engine::graphics

The graphics module handles all drawing operations for any type implementing the Drawable trait.

Modules

buffer
TODO: Partially implemented wrapper for Vulkano buffers.

camera
The camera module holds the different camera options and helper functions for creating and manipulating views.

context
The main Vulkan interface, holds backend components and contextual information such as device, queue, and swapchain information.

error
Holds all graphics error enums.

image
TODO: A module dedicated to images, used for textures and other image related things.

material
TODO: Material is a module that will hold information about how to correctly render a specific object.

shader
The shader module defines types, traits, and structs to abstract complex operations that involve shaders. This module has a lot of intense types from Vulkano wrapped in less scary interfaces that are not as troublesome to deal with

Structs

DescriptorBuilder

DrawInfo

InstanceData

PipelineData

Rect

Vertex

Enums

BlendMode

Transform

Traits

Drawable