Struct ledge_engine::graphics::context::GraphicsContext

pub struct GraphicsContext {
    pub queue: Arc<Queue>,
    pub surface: Arc<Surface<Window>>,
    pub device: Arc<Device>,
    pub swapchain: Arc<Swapchain<Window>>,
    pub sampler: Arc<Sampler>,
    pub framebuffers: Vec<Arc<dyn FramebufferAbstract + Send + Sync>>,
    pub render_pass: Arc<dyn RenderPassAbstract + Send + Sync>,
    pub dynamic_state: DynamicState,
    pub image_num: usize,
    pub acquire_future: Option<SwapchainAcquireFuture<Window>>,
    pub recreate_swapchain: bool,
    pub previous_frame_end: Option<Box<dyn GpuFuture>>,
    pub command_buffer: Option<AutoCommandBufferBuilder<StandardCommandPoolBuilder>>,
    pub pipe_data: Arc<PipelineData>,
}

This is the context from which the graphics components gets all of its information about the physical device and the presentation area. It serves as the Vulkano abstraction, which intern interfaces with the Vulkan API.

Examples

use winit::{
    event_loop::{ControlFlow},
    event::{Event, WindowEvent}
};
use ledge_engine::graphics::context::GraphicsContext;
use ledge_engine::conf::Conf;
 
fn main() {
    let (mut context, event_loop) = GraphicsContext::new(Conf::default());

    event_loop.run(move |event, _, control_flow| {
        let now = std::time::Instant::now();

        match event {
            Event::WindowEvent { event, .. } => match event {
                WindowEvent::CloseRequested => {
                    *control_flow = ControlFlow::Exit;
                },
                WindowEvent::Resized(_) => {
                    context.recreate_swapchain = true;
                },
                _ => {},
            },
            Event::MainEventsCleared => { 
                context.create_command_buffer();
 
                // buffer updates
 
                context.begin_frame();
 
                // draw commands
 
                context.present();
 
                // without using timer you have to manually control the frame time.
                let mut sleep_time: f64 = 0.016 - now.elapsed().as_secs_f64();
                if sleep_time < 0.0 {
                    sleep_time = 0.0
                }
                std::thread::sleep(std::time::Duration::from_secs_f64(sleep_time));
                print!("{:.2}\r", now.elapsed().as_secs_f32() * 1000.0);
            },
            _ => {}
        }
    });
}

Fields
queue: Arc<Queue>surface: Arc<Surface<Window>>device: Arc<Device>swapchain: Arc<Swapchain<Window>>sampler: Arc<Sampler>framebuffers: Vec<Arc<dyn FramebufferAbstract + Send + Sync>>render_pass: Arc<dyn RenderPassAbstract + Send + Sync>dynamic_state: DynamicStateimage_num: usizeacquire_future: Option<SwapchainAcquireFuture<Window>>recreate_swapchain: boolprevious_frame_end: Option<Box<dyn GpuFuture>>command_buffer: Option<AutoCommandBufferBuilder<StandardCommandPoolBuilder>>pipe_data: Arc<PipelineData>
Implementations

impl GraphicsContext

pub fn new(conf: Conf) -> (Self, EventLoop<()>)

pub fn create_command_buffer(&mut self)

Due to the nature of the command buffer and the safety requirements Vulkano tries to meet the command buffer is recreated every frame.

pub fn begin_frame(&mut self)

Handles setup of a new frame, called when the graphics pipeline is first created and at the end of every frame to start the next one.

This is necessary because the swapchain could be out of date, as well as updating the image_num, optimality, and the swapcahin future.

pub fn draw(
&mut self,
vertices: Arc<dyn BufferAccess + Send + Sync>,
shader_handle: Arc<dyn ShaderHandle>,
descriptor: Arc<dyn DescriptorSet + Send + Sync>
)

Interacts with the given shader handle (which by default is a ledge_engine::graphics::shader::ShaderProgram) to use that specific shader to draw the vertex buffer to the screen.

pub fn present(&mut self)

This function submits the command buffer to the queue and fences the operation, storing a future refering to the operation.

This function must be run once at the end of all updates and draw calls in order for the frame to be sumbitted. The context will panic if this is not called once per frame.